Refactor into hpp and cpp files
