+++ include/utility.hpp
@@ -1,18 +1,15 @@
-#ifndef UTILITY_H
-#define UTILITY_H
+#ifndef UTILITY
+#define UTILITY
 
 #include <unordered_set>
 #include <vector>
-// #include "node.h"
 
 class Node;
 
-using std::unordered_set;
-using std::vector;
-
+// TODO: Change these to shared_ptrs
 typedef std::pair<int, int> Pair;
-typedef unordered_set<Node *> node_set;
-typedef vector<Node *> node_vect;
+typedef std::unordered_set<Node *> node_set;
+typedef std::vector<Node *> node_vect;
 
 std::ostream &operator<<(std::ostream &os, const Pair &value)
 {
@@ -38,4 +35,4 @@ std::ostream &operator<<(std::ostream &os, const node_vect &value)
     return os;
 }
 
-#endif // UTILITY_H
+#endif // UTILITY
